+++ lib/input/constructor.go
@@ -80,6 +80,7 @@ const (
 	TypeAMQP            = "amqp"
 	TypeAMQP09          = "amqp_0_9"
 	TypeAMQP1           = "amqp_1"
+	TypeAWSKinesis      = "aws_kinesis"
 	TypeAWSS3           = "aws_s3"
 	TypeBloblang        = "bloblang"
 	TypeBroker          = "broker"
@@ -128,6 +129,7 @@ type Config struct {
 	AMQP            reader.AMQPConfig            `json:"amqp" yaml:"amqp"`
 	AMQP09          reader.AMQP09Config          `json:"amqp_0_9" yaml:"amqp_0_9"`
 	AMQP1           reader.AMQP1Config           `json:"amqp_1" yaml:"amqp_1"`
+	AWSKinesis      AWSKinesisConfig             `json:"aws_kinesis" yaml:"aws_kinesis"`
 	AWSS3           AWSS3Config                  `json:"aws_s3" yaml:"aws_s3"`
 	Bloblang        BloblangConfig               `json:"bloblang" yaml:"bloblang"`
 	Broker          BrokerConfig                 `json:"broker" yaml:"broker"`
@@ -177,6 +179,7 @@ func NewConfig() Config {
 		AMQP:            reader.NewAMQPConfig(),
 		AMQP09:          reader.NewAMQP09Config(),
 		AMQP1:           reader.NewAMQP1Config(),
+		AWSKinesis:      NewAWSKinesisConfig(),
 		AWSS3:           NewAWSS3Config(),
 		Bloblang:        NewBloblangConfig(),
 		Broker:          NewBrokerConfig(),